+++ experiment/experiment.go
@@ -7,6 +7,7 @@ package experiment
 import (
 	"github.com/lawrencewoodman/ddataset"
 	"github.com/vlifesystems/rhkit/aggregators"
+	"github.com/vlifesystems/rhkit/assessment"
 	"github.com/vlifesystems/rhkit/goal"
 	"github.com/vlifesystems/rhkit/internal"
 	"github.com/vlifesystems/rhkit/rule"
@@ -18,13 +19,8 @@ type ExperimentDesc struct {
 	RuleComplexity rule.Complexity
 	Aggregators    []*aggregators.Desc
 	Goals          []string
-	SortOrder      []*SortDesc
+	SortOrder      []assessment.SortDesc
 	Rules          []string
-}
-
-type SortDesc struct {
-	AggregatorName string
-	Direction      string
 }
 
 type Experiment struct {
@@ -33,27 +29,8 @@ type Experiment struct {
 	RuleComplexity rule.Complexity
 	Aggregators    []aggregators.Spec
 	Goals          []*goal.Goal
-	SortOrder      []SortField
+	SortOrder      []assessment.SortOrder
 	Rules          []rule.Rule
-}
-
-type SortField struct {
-	Field     string
-	Direction direction
-}
-
-type direction int
-
-const (
-	ASCENDING direction = iota
-	DESCENDING
-)
-
-func (d direction) String() string {
-	if d == ASCENDING {
-		return "ascending"
-	}
-	return "descending"
 }
 
 // Create a new Experiment from the description
@@ -71,7 +48,7 @@ func New(e *ExperimentDesc) (*Experiment, error) {
 		return nil, err
 	}
 
-	sortOrder, err := makeSortOrder(e.SortOrder)
+	sortOrder, err := assessment.MakeSortOrders(aggregators, e.SortOrder)
 	if err != nil {
 		return nil, err
 	}
@@ -93,37 +70,8 @@ func New(e *ExperimentDesc) (*Experiment, error) {
 }
 
 func checkExperimentDescValid(e *ExperimentDesc) error {
-	if err := checkSortDescsValid(e); err != nil {
-		return err
-	}
 	if err := checkRuleFieldsValid(e); err != nil {
 		return err
-	}
-	return nil
-}
-
-func checkSortDescsValid(e *ExperimentDesc) error {
-	for _, sortDesc := range e.SortOrder {
-		if sortDesc.Direction != "ascending" && sortDesc.Direction != "descending" {
-			return &InvalidSortDirectionError{
-				sortDesc.AggregatorName,
-				sortDesc.Direction,
-			}
-		}
-		sortName := sortDesc.AggregatorName
-		nameFound := false
-		for _, aggregator := range e.Aggregators {
-			if aggregator.Name == sortName {
-				nameFound = true
-				break
-			}
-		}
-		if !nameFound &&
-			sortName != "percentMatches" &&
-			sortName != "numMatches" &&
-			sortName != "goalsScore" {
-			return InvalidSortFieldError(sortName)
-		}
 	}
 	return nil
 }
@@ -142,19 +90,4 @@ func checkRuleFieldsValid(e *ExperimentDesc) error {
 		}
 	}
 	return nil
-}
-
-func makeSortOrder(eSortOrder []*SortDesc) ([]SortField, error) {
-	r := make([]SortField, len(eSortOrder))
-	for i, eSortField := range eSortOrder {
-		field := eSortField.AggregatorName
-		direction := eSortField.Direction
-		// TODO: Make case insensitive
-		if direction == "ascending" {
-			r[i] = SortField{field, ASCENDING}
-		} else {
-			r[i] = SortField{field, DESCENDING}
-		}
-	}
-	return r, nil
 }